Refactored a few things with integrations. Ran in to an issue with integrations being ran prior to the loading of the modules, so provided a way to defer the initialization of some processes if they are dependent upon the modules. Minor formatting to remain consistent.
