+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/economies/SaneEconomy.java
@@ -53,8 +53,7 @@ public class SaneEconomy
     }
 
 	@Override
-	public String getPluginSourceURL()
-	{
+	public String getPluginSourceURL() {
 		return "https://www.spigotmc.org/resources/saneeconomy-simple-but-featureful-economy.26223/";
 	}
 }
